# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cab Booking Backend Service in Go.",
        "title": "Cab Booking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/bookings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passenger"
                ],
                "summary": "Request a new ride",
                "parameters": [
                    {
                        "description": "Ride Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.CreateBookingResponse"
                        }
                    }
                }
            }
        },
        "/driver/bookings/{bookingId}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Accept a requested ride",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.DriverActionResponse"
                        }
                    }
                }
            }
        },
        "/driver/bookings/{bookingId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Cancel a ride (Driver)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.DriverActionResponse"
                        }
                    }
                }
            }
        },
        "/driver/bookings/{bookingId}/end": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Complete a ride",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.DriverActionResponse"
                        }
                    }
                }
            }
        },
        "/driver/bookings/{bookingId}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Driver"
                ],
                "summary": "Start a ride (Verify OTP)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "OTP JSON: {\\",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.DriverActionResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get JWT token",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1.LoginResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CabBookingService_internal_models.BookingStatus": {
            "type": "string",
            "enum": [
                "REQUESTED",
                "ACCEPTED",
                "STARTED",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "BookingStatusRequested",
                "BookingStatusAccepted",
                "BookingStatusStarted",
                "BookingStatusCompleted",
                "BookingStatusCancelled"
            ]
        },
        "internal_controllers_v1.CreateBookingRequest": {
            "type": "object",
            "properties": {
                "dropoff_lat": {
                    "type": "number"
                },
                "dropoff_lon": {
                    "type": "number"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lon": {
                    "type": "number"
                }
            }
        },
        "internal_controllers_v1.CreateBookingResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dropoff_lat": {
                    "type": "number"
                },
                "dropoff_lon": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lon": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/CabBookingService_internal_models.BookingStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1.DriverActionResponse": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1.RegisterPassengerRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1.RegisterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}